Add quirk to force port in Host header
Some devices rely on finding the port in the Host HTTP header to
construct their response to the /eSCL/ScanJobs request.  When the port
isn't included, the steps up to the scan job creation work, but an
invalid redirect will be sent in the response to /eSCL/ScanJobs and then
the scan job will never finish.

Add a new quirk that forces the port to always be included in the Host
header.  This is known to affect some Epson models, and Epson has
indicated that this should be safe to use with all of their devices, so
enable it for eSCL whenever the manufacturer is Epson.  The quirk itself
is generic so that it can easily be enabled for more devices in the
future if needed.

Tested on a ET-3750.
Fixes #136.