Makefile: abstract ar to $(AR) and add default target
Build systems need to be able to use a custom `ar` binary.

This also adds a "default" target, which is run when `make` is called
with no arguments. This keeps the current behavior and calls ctags and
such. Then, the "all" target can be used by distros to build this
without generating ctags.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>