Fixed build with recent versions of libxml2 (fixes #305)
libxml2 recently changed a prototype of the xmlStructuredErrorFunc callback,
adding const modifier to the second parameter.

Adding an explicit cast seems to fix the problem (but I will not be surprised, if
this fix will not work equally well on all distros).

Found for a first time and fixed on Arch linux on about 2 Dec 2023 (Arch is rolling
release, so there is no explicit release number)