ESCL: workaround for HP Deskjet 3520 series over IPP over USB
This device returns Location: http://CN26F178X605SZ.03f011b0.hpLedmUSB/...
instead of expected http://localhost/... as response to the ScanJobs
request (please notice the hostname).

We actually used to ignore hostname in the ScanJobs response in
assumption that device may not have a strong knowledge of its own
hostname, but a while ago this logic was occasionally broken.
This patch restores that behavior.