WSD: sca:ScannerDescription requested in sca:GetScannerElementsRequest
The only reason for this change is to make this information visible
in the log.

I need it to investigate possible workarounds for #334. I'm trying to
find some way to detect Ricoh Aficio MP 201 ADF, to make workaround
as specific as possible, so it will not not be applied to other devices.