WSD: fix for ADF scan on RICOH Aficio MP 201 (see #334)
This device doesn't properly handle sca:ImagesToTransfer if set to "0".

Instead of scanning all available sheets from feeder, as defined by
protocol, it only scans a single page.

The workaround is to set sca:ImagesToTransfer to some big arbitrary number.

Please note, if device is configured manually, we don't have a reliable
indication of this model name. From another hand, I'm not sure how other
devices will behave with this setting, so don't want to apply it to
all.

So this solution is technical compromise. It will only work when device
is used with zeroconf.