WSD: Add content type selection
Handle the WSD device settings element `ContentTypesSupported`, present
its values as an option, and pass the selection to the scan job as
`ContentType`.

The actual identifiers in code are made to match eSCL's "scan intent"
setting, except for content type values exclusive to WSD. The setting is
presented as `scan-intent`.

On some devices, this setting affects how colour is processed, and the
default mode when not specified can result in inaccurate colours.