Added ID_SCANINTENT_UNSET value for the 'sane-intent'
If ID_SCANINTENT_UNSET is selected, the intent is not sent to the device.
This is compatible with the previous version of the driver.

This compatibility is important because, sooner or later, we will
encounter a device that supports some intents but fails to scan when one
of them is actually selected. Having an explicit unset option provides
users with a reasonable fallback.

ID_SCANINTENT_UNSET is the default value for the 'sane-intent' option. It
helps avoid introducing unexpected changes in scanning quality for users
who have been using sane-airscan for some time.