HTTP timeout for "short" requests extended from 5s to 20s
Under certain circumstances, when the Kyocera ECOSYS M2040dn is
connected via ipp-usb, it can take up to 12 seconds to respond to the
eSCL/ScannerStatus after a successful scan.

If we allow the scanner this additional time, it continues to operate
correctly. However, with the default timeouts, this delay can cause
scanning to fail.

This issue is difficult to reproduce; it only occurs with XSane when
scanning from the flatbed.

It may also affect other scanners, but so far, no one has reported a
similar issue.

Everything else appears to be functioning correctly on both the
sane-airscan and ipp-usb sides.

I will continue to investigate this problem on the ipp-usb side. For
now, I have increased the HTTP timeouts with some reasonable reserve for
the "short" operations, such as requesting the scanner status and
similar.

This change is safe; it does not impact scanning when everything is
functioning normally, but it may extend the time required for failure
detection.